Added acceleration display updates every heartbeat from spatial accelerometer.

+++ flightsoft/trunk/capture/Message/AccelerometerStatusMessage.cs
@@ -9,7 +9,7 @@ namespace uGCapture
     public class AccelerometerStatusMessage: StatusMessage
     {
         public float accel = 0;
-        public AccelerometerStatusMessage(Receiver s, Status nstate, ErrStr mes = 0, int accl=0)
+        public AccelerometerStatusMessage(Receiver s, Status nstate, ErrStr mes = 0, double accl=0)
             : base(s, nstate, mes)
         {
             accel = accl;
